Make projects use the github master version of the nixexpr functions
Currently new haskell-overridez projects
1) load the support functions using import-from-derivation, importing
  the support functions directly from github
2) the script makes this use the latest commit on github

This changes 2) to use only the latest master version on github.

While this is not as good as using specific versions, it is a big improvement.
The master branch is not the default branch (develop), and only has a latest
"released" version, i.e, it is a lot more stable.

It also means that the script installation instructions and the project
initialization now use the same version.