Step 1 - add new test project
A new test project is required.

The tests need to make use of the executable build by the main project. With
just cabal, this is possible using a test-target inside the same project via
referring to the executable using build-depends

However, normally, nix won't install the built executable until the tests for
the project have run, so nix builds will not work.